Moving rowwise_gradients to tf_wraps
